+++ test/assets/scripts/test.js
@@ -1,0 +1,5 @@
+if (typeof describe == 'function') {
+	return;
+}
+
+console.log('This is a test script');